remove tokenizer package and rename to TeragrepTokenizerStep. remove unused interface
